Lazy load auth and ai tabs only when ext is enabled + lazy load schema graph
