+++ shared/studio/components/lazyTabs/lazyTabs.module.scss
@@ -1,0 +1,83 @@
+@import "@edgedb/common/mixins.scss";
+
+.tabWrapper {
+  position: relative;
+  flex-grow: 1;
+  background-color: var(--app_panel_background);
+  border-radius: 12px;
+  box-shadow: 0 0 8px rgba(0, 0, 0, 0.04), 0 0 4px rgba(0, 0, 0, 0.06);
+  display: flex;
+  min-width: 0;
+  min-height: 0;
+  font-family: "Roboto Flex Variable", sans-serif;
+
+  @include isMobile {
+    flex-direction: column;
+    border-radius: 0;
+  }
+
+  @include darkTheme {
+    box-shadow: 0px 0px 8px 0px rgba(0, 0, 0, 0.2),
+      0px 0px 4px 0px rgba(0, 0, 0, 0.3);
+  }
+}
+
+.loadingSchema {
+  flex-grow: 1;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  font-style: italic;
+}
+
+.fallbackSpinner {
+  margin: auto;
+}
+
+.extDisabled {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: center;
+  text-align: center;
+  margin: auto;
+  font-family: "Roboto Flex Variable", sans-serif;
+  color: #4d4d4d;
+
+  h2 {
+    color: inherit;
+    font-size: 18px;
+    font-style: normal;
+    font-weight: 500;
+    line-height: 26px;
+    margin: 0;
+  }
+
+  p {
+    margin: 8px 0;
+  }
+
+  pre {
+    background: #eee;
+    margin: 12px;
+    padding: 12px 16px;
+    border-radius: 6px;
+    user-select: text;
+
+    @include darkTheme {
+      background: #383838;
+    }
+  }
+
+  a {
+    color: #b044c2;
+
+    @include darkTheme {
+      color: #ac86f6;
+    }
+  }
+
+  @include darkTheme {
+    color: #ccc;
+  }
+}
