+++ shared/studio/tabs/ai/aiAdmin.module.scss
@@ -1,81 +1,5 @@
 @import "@edgedb/common/mixins.scss";
 
-.aiAdmin {
-  position: relative;
-  flex-grow: 1;
-  border-radius: 12px;
-  background: #fbfbfb;
-  color: #4d4d4d;
-  box-shadow: 0px 0px 8px 0px rgba(0, 0, 0, 0.04),
-    0px 0px 4px 0px rgba(0, 0, 0, 0.06);
-  display: flex;
-  justify-content: center;
-  gap: 32px;
-  min-width: 0;
-  min-height: 0;
-  font-family: "Roboto Flex Variable", sans-serif;
-
-  h2 {
-    color: inherit;
-    font-size: 18px;
-    font-style: normal;
-    font-weight: 500;
-    line-height: 26px;
-    margin: 0;
-  }
-
-  @include darkTheme {
-    color: #ccc;
-    background: var(--Grey14, #242424);
-    box-shadow: 0px 0px 8px 0px rgba(0, 0, 0, 0.2),
-      0px 0px 4px 0px rgba(0, 0, 0, 0.3);
-  }
-
-  @include isMobile {
-    border-radius: 0;
-  }
-}
-
-.loadingSchema {
-  flex-grow: 1;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  font-style: italic;
-}
-
-.extDisabled {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
-  text-align: center;
-
-  p {
-    margin: 8px 0;
-  }
-
-  pre {
-    background: #eee;
-    margin: 12px;
-    padding: 12px 16px;
-    border-radius: 6px;
-    user-select: text;
-
-    @include darkTheme {
-      background: #383838;
-    }
-  }
-
-  a {
-    color: #b044c2;
-
-    @include darkTheme {
-      color: #ac86f6;
-    }
-  }
-}
-
 .mainLayout {
   flex-grow: 1;
   display: grid;
@@ -93,6 +17,15 @@
   grid-column: 2 / span 2;
   min-height: 0;
   min-width: 0;
+
+  h2 {
+    color: inherit;
+    font-size: 18px;
+    font-style: normal;
+    font-weight: 500;
+    line-height: 26px;
+    margin: 0;
+  }
 
   @include isMobile {
     grid-column: 1;
@@ -168,8 +101,10 @@
     justify-self: stretch;
     flex-direction: row;
     justify-content: center;
+    background: var(--header_background);
     padding: 4px 8px;
     box-shadow: 0 2px 6px rgba(0, 0, 0, 0.04), 0 2px 3px rgba(0, 0, 0, 0.03);
+    z-index: 2;
 
     .tab {
       display: flex;
@@ -1079,7 +1014,7 @@
   &:before {
     content: "";
     position: absolute;
-    background: #fbfbfb;
+    background: var(--app_panel_background);
     top: -16px;
     left: -1px;
     right: -1px;
