+++ shared/studio/tabs/auth/index.tsx
@@ -1,74 +1,34 @@
-import {useEffect} from "react";
+import {lazy, Suspense} from "react";
 import {observer} from "mobx-react-lite";
 
-import cn from "@edgedb/common/utils/classNames";
+import CodeBlock from "@edgedb/common/ui/codeBlock";
+import Spinner from "@edgedb/common/ui/spinner";
 
-import styles from "./authAdmin.module.scss";
+import {useDatabaseState} from "../../state";
 
 import {DatabaseTabSpec} from "../../components/databasePage";
-
-import {useTabState} from "../../state";
-
 import {TabAuthIcon} from "../../icons";
-import {AuthAdminState, _providersInfo} from "./state";
 
-import CodeBlock from "@edgedb/common/ui/codeBlock";
-import {CustomScrollbars} from "@edgedb/common/ui/customScrollbar";
-import {WebhooksTab} from "./webhooks";
+import styles from "../../components/lazyTabs/lazyTabs.module.scss";
 
-import {PanelTabs} from "@edgedb/common/newui/panelTabs";
-import {ConfigTab} from "./config";
-import {SMTPConfigTab} from "./smtp";
-import {ProvidersTab} from "./providers";
+const AuthAdmin = lazy(() => import("./auth"));
 
-export const AuthAdmin = observer(function AuthAdmin() {
-  const state = useTabState(AuthAdminState);
+const AuthAdminLoader = observer(function AuthAdminLoader() {
+  const db = useDatabaseState();
 
-  useEffect(() => {
-    if (state.extEnabled) {
-      state.refreshConfig();
-    }
-  }, [state.extEnabled]);
+  const extEnabled =
+    db.schemaData?.extensions.some((ext) => ext.name === "auth") ?? null;
 
   return (
-    <div
-      className={cn(styles.authAdmin, {
-        [styles.loaded]: state.extEnabled === true,
-      })}
-    >
-      {state.extEnabled === null ? (
+    <div className={styles.tabWrapper}>
+      {extEnabled === null ? (
         <div className={styles.loadingSchema}>Loading schema...</div>
-      ) : state.extEnabled ? (
-        <CustomScrollbars
-          className={styles.scrollWrapper}
-          innerClass={styles.scrollContent}
+      ) : extEnabled ? (
+        <Suspense
+          fallback={<Spinner className={styles.fallbackSpinner} size={20} />}
         >
-          <div className={styles.scrollContent}>
-            <PanelTabs
-              className={styles.tabs}
-              tabs={[
-                {id: "config", label: "Config"},
-                ...(state.hasWebhooksSchema
-                  ? [{id: "webhooks", label: "Webhooks"} as const]
-                  : []),
-                {id: "providers", label: "Providers / UI"},
-                {id: "smtp", label: "SMTP"},
-              ]}
-              selectedTabId={state.selectedTab}
-              setSelectedTabId={(id) => state.setSelectedTab(id)}
-            />
-
-            {state.selectedTab === "config" ? (
-              <ConfigTab />
-            ) : state.selectedTab === "providers" ? (
-              <ProvidersTab />
-            ) : state.selectedTab === "webhooks" ? (
-              <WebhooksTab />
-            ) : (
-              <SMTPConfigTab />
-            )}
-          </div>
-        </CustomScrollbars>
+          <AuthAdmin />
+        </Suspense>
       ) : (
         <div className={styles.extDisabled}>
           <h2>The auth extension is not enabled</h2>
@@ -91,6 +51,5 @@ export const authAdminTabSpec: DatabaseTabSpec = {
   label: "Auth Admin",
   icon: (active) => <TabAuthIcon active={active} />,
   usesSessionState: false,
-  element: <AuthAdmin />,
-  state: AuthAdminState,
+  element: <AuthAdminLoader />,
 };
