+++ shared/studio/tabs/schema/index.tsx
@@ -1,3 +1,4 @@
+import {lazy, Suspense} from "react";
 import {observer} from "mobx-react-lite";
 
 import cn from "@edgedb/common/utils/classNames";
@@ -9,14 +10,6 @@ import styles from "./schema.module.scss";
 
 import {useTabState} from "../../state";
 import {Schema, SchemaViewType} from "./state";
-
-import {
-  SchemaGraph,
-  schemaContext,
-  useDebugState,
-  DebugControls,
-  SchemaMinimap,
-} from "@edgedb/schema-graph";
 
 import {DatabaseTabSpec} from "../../components/databasePage";
 
@@ -31,6 +24,18 @@ import {SchemaTextView} from "./textView";
 import {useIsMobile} from "@edgedb/common/hooks/useMobile";
 import {LabelsSwitch, switchState} from "@edgedb/common/ui/switch";
 
+const _SchemaGraphView = lazy(() => import("./graphView"));
+
+function SchemaGraphView({state}: {state: Schema}) {
+  return (
+    <div className={styles.schemaGraphView}>
+      <Suspense fallback={null}>
+        <_SchemaGraphView state={state} />
+      </Suspense>
+    </div>
+  );
+}
+
 export const SchemaView = observer(function SchemaView() {
   const schemaState = useTabState(Schema);
   const isMobile = useIsMobile();
@@ -38,73 +43,67 @@ export const SchemaView = observer(function SchemaView() {
   const isGraphView = schemaState.viewType === SchemaViewType.Graph;
 
   return (
-    <schemaContext.Provider value={schemaState.schemaState}>
-      <div className={cn(styles.schema)}>
-        {schemaState.viewType === SchemaViewType.Text ? (
-          <SchemaTextView />
-        ) : null}
-        {isGraphView ? <SchemaGraphView /> : null}
-        {schemaState.viewType === SchemaViewType.TextGraph ? (
-          <SplitView
-            views={[<SchemaTextView />, <SchemaGraphView />]}
-            state={schemaState.splitView}
-            minViewSize={20}
-          />
-        ) : null}
+    <div className={cn(styles.schema)}>
+      {schemaState.viewType === SchemaViewType.Text ? (
+        <SchemaTextView />
+      ) : null}
+      {isGraphView ? <SchemaGraphView state={schemaState} /> : null}
+      {schemaState.viewType === SchemaViewType.TextGraph ? (
+        <SplitView
+          views={[<SchemaTextView />, <SchemaGraphView state={schemaState} />]}
+          state={schemaState.splitView}
+          minViewSize={20}
+        />
+      ) : null}
 
-        {isMobile ? (
-          <LabelsSwitch
-            className={styles.viewSwitch}
-            labels={["Text", "Graph"]}
-            value={isGraphView ? switchState.right : switchState.left}
-            onChange={() =>
-              schemaState.setViewType(
-                isGraphView ? SchemaViewType.Text : SchemaViewType.Graph
-              )
-            }
-          />
-        ) : (
-          <div className={cn(styles.toolbar)}>
-            <div className={styles.switcherLabel}>View Layout</div>
-            <div className={styles.viewSwitcher}>
-              <SwitcherButton
-                items={[
-                  {
-                    id: SchemaViewType.Text,
-                    label: "Text",
-                    icon: (
-                      <SchemaViewTextIcon
-                        className={styles.viewSwitcherIcon}
-                      />
-                    ),
-                  },
-                  {
-                    id: SchemaViewType.Graph,
-                    label: "Graph",
-                    icon: (
-                      <SchemaViewGraphIcon
-                        className={styles.viewSwitcherIcon}
-                      />
-                    ),
-                  },
-                  {
-                    id: SchemaViewType.TextGraph,
-                    label: "Text/Graph",
-                    icon: (
-                      <SchemaViewTextGraphIcon
-                        className={styles.viewSwitcherIcon}
-                      />
-                    ),
-                  },
-                ]}
-                selected={schemaState.viewType}
-                onChange={(type) => schemaState.setViewType(type)}
-              />
-            </div>
+      {isMobile ? (
+        <LabelsSwitch
+          className={styles.viewSwitch}
+          labels={["Text", "Graph"]}
+          value={isGraphView ? switchState.right : switchState.left}
+          onChange={() =>
+            schemaState.setViewType(
+              isGraphView ? SchemaViewType.Text : SchemaViewType.Graph
+            )
+          }
+        />
+      ) : (
+        <div className={cn(styles.toolbar)}>
+          <div className={styles.switcherLabel}>View Layout</div>
+          <div className={styles.viewSwitcher}>
+            <SwitcherButton
+              items={[
+                {
+                  id: SchemaViewType.Text,
+                  label: "Text",
+                  icon: (
+                    <SchemaViewTextIcon className={styles.viewSwitcherIcon} />
+                  ),
+                },
+                {
+                  id: SchemaViewType.Graph,
+                  label: "Graph",
+                  icon: (
+                    <SchemaViewGraphIcon className={styles.viewSwitcherIcon} />
+                  ),
+                },
+                {
+                  id: SchemaViewType.TextGraph,
+                  label: "Text/Graph",
+                  icon: (
+                    <SchemaViewTextGraphIcon
+                      className={styles.viewSwitcherIcon}
+                    />
+                  ),
+                },
+              ]}
+              selected={schemaState.viewType}
+              onChange={(type) => schemaState.setViewType(type)}
+            />
           </div>
-        )}
-      </div>
-    </schemaContext.Provider>
+        </div>
+      )}
+    </div>
   );
 });
 
@@ -117,22 +116,3 @@ export const schemaTabSpec: DatabaseTabSpec = {
   state: Schema,
   element: <SchemaView />,
 };
-
-const SchemaGraphView = observer(function SchemaGraphView() {
-  const schemaState = useTabState(Schema);
-
-  const debugState = useDebugState();
-
-  return (
-    <div className={styles.schemaGraphView}>
-      {process.env.NODE_ENV === "development" ? (
-        <DebugControls
-          debugState={debugState}
-          schemaState={schemaState.schemaState}
-        />
-      ) : null}
-      <SchemaGraph debug={debugState[0]} />
-      <SchemaMinimap className={styles.minimap} />
-    </div>
-  );
-});
